re-use single example form
