+++ src/routes/+page.svelte
@@ -1,3 +1,2 @@
-<a href="/fields">Fields</a>
 <a href="/form">Form</a>
 <a href="/ssr">SSR</a>
